+++ lib/src/state_machine/mp_th2_file_edit_state_machine/types/mp_button_type.dart
@@ -11,6 +11,8 @@ enum MPButtonType {
   addXTherionImageInsertConfig,
   changeScrap,
   chooseTHConfigAndRunTherion,
+  convertLineSegmentsToBezier,
+  convertLineSegmentsToStraight,
   copyElements,
   createMapConnectionLines,
   cutElements,